+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Added option `--fix-serie` to `ci check-versions-git` command.
 - Added command `odoo-helper odoo recompute-menu`
 - Added command `odoo-helper odoo clean-compiled-assets`
+- Added `--no-backup` option to `odoo-helper install reinstall-venv`
 
 
 ### Changed